chore: further multi platform preparation

+++ aeronet/tech/include/aeronet/raw-bytes.hpp
@@ -9,7 +9,7 @@
 namespace aeronet {
 
 // A byte buffer.
-using RawBytes = RawBytesBase<std::byte, std::span<const std::byte>, std::size_t>;
+using RawBytes = RawBytesBase<std::byte, std::span<const std::byte>, std::uint64_t>;
 
 // A byte buffer with 32-bit size type.
 using RawBytes32 = RawBytesBase<std::byte, std::span<const std::byte>, std::uint32_t>;
